+++ accounting/RELEASES.md
@@ -159,3 +159,13 @@ Based on:
 - [typescript v0.7.4] accounting
 ### Releases
 - [NPM v0.7.4] https://www.npmjs.com/package/@speakeasy-sdks/accounting/v/0.7.4 - accounting
+
+## 2024-03-18 01:23:11
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.209.3 (2.281.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.7.5] accounting
+### Releases
+- [NPM v0.7.5] https://www.npmjs.com/package/@speakeasy-sdks/accounting/v/0.7.5 - accounting